Add files via upload
A maze is in the form of a 2D matrix in which some
cells/blocks are blocked. One of the cells is termed as a
source cell, from where we have to start. And another one of
them is termed as a destination cell, where we have to reach.
We have to find a path from the source to the destination
without moving into any of the blocked cell.